+++ packages/expo-camera/android/src/main/java/expo/modules/camera/CameraViewModule.kt
@@ -3,6 +3,9 @@ package expo.modules.camera
 import android.Manifest
 import android.graphics.Bitmap
 import android.util.Log
+import com.google.mlkit.vision.codescanner.GmsBarcodeScanner
+import com.google.mlkit.vision.codescanner.GmsBarcodeScannerOptions
+import com.google.mlkit.vision.codescanner.GmsBarcodeScanning
 import expo.modules.camera.analyzers.BarCodeScannerResultSerializer
 import expo.modules.camera.analyzers.MLKitBarCodeScanner
 import expo.modules.camera.records.BarcodeSettings
@@ -45,6 +48,10 @@ class CameraViewModule : Module() {
     Name("ExpoCamera")
 
     Events("onModernBarcodeScanned")
+
+    Property("isModernBarcodeScannerAvailable") {
+      true
+    }
 
     AsyncFunction("requestCameraPermissionsAsync") { promise: Promise ->
       Permissions.askForPermissionsWithPermissionsManager(
@@ -100,6 +107,26 @@ class CameraViewModule : Module() {
           }
         }
       )
+    }
+
+    AsyncFunction("launchScanner") { settings: BarcodeSettings ->
+      val reactContext = appContext.reactContext
+        ?: throw Exceptions.ReactContextLost()
+      val options = GmsBarcodeScannerOptions.Builder()
+        .setBarcodeFormats(
+          settings.barcodeTypes.first().mapToBarcode(),
+          *settings.barcodeTypes.drop(1).map { it.mapToBarcode() }.toIntArray())
+        .build()
+
+      val scanner = GmsBarcodeScanning.getClient(reactContext, options)
+      scanner.startScan()
+        .addOnSuccessListener { barcode ->
+          val result = BarCodeScannerResultSerializer.parseBarcodeScanningResult(barcode)
+          sendEvent("onModernBarcodeScanned", BarCodeScannerResultSerializer.toBundle(result, 1.0f))
+        }
+        .addOnFailureListener {
+          throw CameraExceptions.BarcodeScanningFailedException()
+        }
     }
 
     OnDestroy {